% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{corr_plot}
\alias{corr_plot}
\title{Plot relationship between p and q
Plot -log10(p) values for quantiles of q}
\usage{
corr_plot(p, q, ylim = c(0, 1.5))
}
\arguments{
\item{p}{p values for principal trait (vector of length n)}

\item{q}{auxillary data values (vector of length n)}

\item{ylim}{y-axis limits (-log10)}
}
\value{
ggplot object
}
\description{
Plot relationship between p and q
Plot -log10(p) values for quantiles of q
}
