% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{log10pv_plot}
\alias{log10pv_plot}
\title{ggplot to visualise fcfdr results (-log10)}
\usage{
log10pv_plot(p, q, v, axis_lim = c(0, 20))
}
\arguments{
\item{p}{p values for principal trait (vector of length n)}

\item{q}{auxillary data values (vector of length n)}

\item{v}{v values from cFDR}

\item{axis_lim}{Optional axis limits}
}
\value{
ggplot object
}
\description{
ggplot to visualise fcfdr results (-log10)
}
