% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_cfdr.R
\name{match_ind_maf}
\alias{match_ind_maf}
\title{Function to downsample independent SNPs to match MAF distribution of whole set.}
\usage{
match_ind_maf(maf, indep_index)
}
\arguments{
\item{maf}{minor allele frequencies of (all) SNPs}

\item{indep_index}{indices of independent SNPs}
}
\value{
indices of independent SNP in chosen in sample
}
\description{
Matches MAF distribution of independent set of SNPs to MAF distribution of whole set of SNPs to avoid MAF-based confounding.
}
\details{
Must supply maf values from the whole data set, not just the independent SNPs.
}
