% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexible_cfdr.R
\name{parameters_in_locfdr}
\alias{parameters_in_locfdr}
\title{parameters_in_locfdr}
\usage{
parameters_in_locfdr(
  p,
  q,
  indep_index,
  res_p = 300,
  res_q = 500,
  maf = NULL,
  check_indep_cor = TRUE,
  enforce_p_q_cor = TRUE
)
}
\arguments{
\item{p}{p values for principal trait (vector of length n)}

\item{q}{continuous auxiliary data values (vector of length n)}

\item{indep_index}{indices of independent SNPs}

\item{res_p}{resolution for p}

\item{res_q}{resolution for q}

\item{maf}{minor allele frequencies for SNPs to which \code{p} and \code{q} relate (optional and used to perform MAF matching)}

\item{check_indep_cor}{check that sign of the correlation between \code{p} and \code{q} is the same in the independent subset as in the whole}

\item{enforce_p_q_cor}{if \code{p} and \code{q} are negatively correlated, flip the sign on \code{q} values}
}
\value{
list of values used as input into \code{locfdr::locfdr} function intrinsically in \code{flexible_cfdr}
}
\description{
parameters_in_locfdr
}
