% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_cfdr.R
\name{per_group_binary_cfdr}
\alias{per_group_binary_cfdr}
\title{Process one fold of data for binary cFDR.}
\usage{
per_group_binary_cfdr(p_loo, q_loo, ps, qs, x)
}
\arguments{
\item{p_loo}{p-values sans current fold}

\item{q_loo}{covariate values sans current fold}

\item{ps}{p-values from current fold}

\item{qs}{covariate values from current fold}

\item{x}{}
}
\description{
Process one fold of data for binary cFDR.
}
