% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T1D_df}
\alias{T1D_df}
\title{T1D Application data}
\format{
A data frame with 114641 rows and 11 variables:
}
\usage{
T1D_df
}
\description{
A dataframe containing the chromosome (chrom), base pair position (pos),
rsid, major (other_allele) and minor (effect_allele) allele, T1D GWAS p-value (p),
binary chromatin accessibility measure (DGF_ENCODE), H3K27ac fold change value
in naive CD4+ T helper cells (Th_H3K27ac), RA GWAS p-value (RA_p), ldak weight, and minor allele frequency (maf) for 114,641 SNPs in the T1D GWAS (https://www.nature.com/articles/ng.3245)
}
\details{
Minor allele frequency is taken from data by a study by Cooper et al.
(https://doi.org/10.1101/120022) where available. Missing values were replaced
by drawing samples from the empirical distribution of MAFs
}
\keyword{datasets}
