% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_cfdr.R
\name{binary_cfdr}
\alias{binary_cfdr}
\title{Function to perform cFDR for binary auxiliary covariates}
\usage{
binary_cfdr(p, q, chr, p.thr = 0.01, digits_highp = 1, digits_lowp = 2)
}
\arguments{
\item{p}{p values for principal trait (vector of length n)}

\item{q}{binary auxillary data values (vector of length n)}

\item{chr}{chromosome each SNP resides (vector of length n)}

\item{p.thr}{p-value threshold for interpolation based on digits_highp and digits_lowp}

\item{digits_highp}{digits for rounding for p >= p.thr}

\item{digits_lowp}{digits for rounding for p < p.thr}
}
\value{
dataframe of p, q and v values
}
\description{
Function to perform cFDR for binary auxiliary covariates
}
