% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_cfdr.R
\name{binary_cfdr}
\alias{binary_cfdr}
\title{Perform cFDR leveraging binary auxiliary covariates}
\usage{
binary_cfdr(p, q, group)
}
\arguments{
\item{p}{p-values for principal trait (vector of length n)}

\item{q}{binary auxiliary data values (vector of length n)}

\item{group}{group membership of each SNP for leave-one-out procedure (vector of length n) (e.g. chromosome number or LD block)}
}
\value{
data.frame of p, q and v values
}
\description{
Perform cFDR leveraging binary auxiliary covariates
}
