% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{corr_plot}
\alias{corr_plot}
\title{Violin plot of p-values for quantiles of q}
\usage{
corr_plot(p, q, ylim = c(0, 1.5))
}
\arguments{
\item{p}{p values for principal trait (vector of length n)}

\item{q}{auxiliary data values (vector of length n)}

\item{ylim}{y-axis limits (-log10)}
}
\value{
ggplot object
}
\description{
Violin plot of p-values for quantiles of q
}
\details{
Can be used to investigate the relationship between p and q

If this shows a non-monotonic relationship then the cFDR framework should not be used

(because e.g. cFDR cannot simultaneously shrink v-values for high p and low p)
}
