% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/df.R
\docType{data}
\name{df}
\alias{df}
\title{Dataframe for T1D analysis}
\format{
a data frame

A data frame with 114641 rows and 11 variables:
}
\usage{
df

df
}
\description{
data.frame object containing columns for chromosome, position (GRCh37),
rsid, effect allele, reference allele, T1D p-value from Onengut-Gumuscu et al. (2015),
binary DGF annotation, continuous H3K27ac fold enrichment value in T helper cells, RA p-value
from Eyre et al. (2012) and LDAK weight for each SNP (row)

A dataframe containing the chromosome (chrom), base pair position (pos),
rsid, major (other_allele) and minor (effect_allele) allele, T1D GWAS p-value (p),
binary chromatin accessibility measure (DGF_ENCODE), H3K27ac fold change value
in naive CD4+ T helper cells (Th_H3K27ac), RA GWAS p-value (RA_p), ldak weight, and minor allele frequency (maf) for 114,641 SNPs in the T1D GWAS (https://www.nature.com/articles/ng.3245)
}
\details{
Minor allele frequency is taken from data by a study by Cooper et al.
(https://doi.org/10.1101/120022) where available. Missing values were replaced
by drawing samples from the uniform distribution on \link{0,0.5}
}
\keyword{datasets}
