% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{log10pv_plot}
\alias{log10pv_plot}
\title{log10pv_plot
ggplot to visualise the results from the re-weighting, coloured by the value of q.}
\usage{
log10pv_plot(output)
}
\arguments{
\item{output}{Output from flexible_cfdr function}
}
\value{
ggplot object
}
\description{
log10pv_plot
ggplot to visualise the results from the re-weighting, coloured by the value of q.
}
