% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{pv_plot}
\alias{pv_plot}
\title{Plot p against v and colour by q}
\usage{
pv_plot(p, q, v, axis_lim = c(0, 1))
}
\arguments{
\item{p}{p values for principal trait (vector of length n)}

\item{q}{auxiliary data values (vector of length n)}

\item{v}{v values from cFDR}

\item{axis_lim}{Optional axis limits}
}
\value{
ggplot object
}
\description{
Plot p against v and colour by q
}
\details{
Can be used to visualise the results from Flexible cFDR
}
